refactored pytorch code significantly
