--- MuyGPyS/examples/torch_demo.py
@@ -1,4 +1,0 @@
-# Copyright 2021-2022 Lawrence Livermore National Security, LLC and other
-# MuyGPyS Project Developers. See the top-level COPYRIGHT file for details.
-#
-# SPDX-License-Identifier: MIT
